+++ src/node/orpc/formatOrpcError.ts
@@ -1,6 +1,8 @@
 import { ORPCError, ValidationError } from "@orpc/server";
 import type { StandardSchemaV1 } from "@standard-schema/spec";
 import { inspect } from "node:util";
+import { DYNAMIC_WORKFLOWS_DISABLED_ERROR_MESSAGE } from "@/node/services/workflows/WorkflowService";
+import { WorkflowArgsValidationError } from "@/node/services/workflows/workflowArgs";
 
 export interface FormattedOrpcError {
   /**
@@ -392,4 +394,14 @@ export function formatOrpcError(error: unknown, interceptorOptions?: unknown): F
       },
     };
   }
+}
+
+export function throwWorkflowOrpcError(error: unknown): never {
+  if (
+    error instanceof WorkflowArgsValidationError ||
+    (error instanceof Error && error.message === DYNAMIC_WORKFLOWS_DISABLED_ERROR_MESSAGE)
+  ) {
+    throw new ORPCError("BAD_REQUEST", { message: error.message });
+  }
+  throw error;
 }
