+++ src/node/services/workspaceService.ts
@@ -201,6 +201,7 @@ import {
   WORKFLOW_RESULT_METADATA_TYPE,
   WORKFLOW_RUN_CARD_DISPLAY_METADATA_TYPE,
   WORKFLOW_TRIGGER_DISPLAY_METADATA_TYPE,
+  buildWorkflowResultContextMessage,
   buildWorkflowRunCardMessage,
   isTerminalWorkflowRunToolOutput,
   isWorkflowRunEmittingToolName,
@@ -1675,6 +1676,89 @@ async function forEachWithConcurrencyLimit<T>(
   });
 
   await Promise.all(workers);
+}
+
+const WORKFLOW_CONTINUATION_RETRY_DELAY_MS = 1_000;
+const WORKSPACE_BUSY_IDLE_ONLY_SEND_MESSAGE = "Workspace is busy; idle-only send was skipped.";
+
+type WorkflowContinuationWorkspaceService = Pick<
+  WorkspaceService,
+  "isWorkflowInvocationCurrent" | "getWorkflowContinuationSendOptions" | "sendMessage"
+>;
+
+export async function sendWorkflowRunTerminalContinuation(
+  service: WorkflowContinuationWorkspaceService,
+  input: {
+    workspaceId: string;
+    rawCommand: string;
+    name: string;
+    runId: string;
+    status: WorkflowRunStatus;
+    result: unknown;
+    run: WorkflowRunRecord;
+    continuationOptions?: SendMessageOptions;
+  }
+): Promise<void> {
+  const commandPrefix = input.rawCommand.split(/\s+/u)[0] ?? input.name;
+  const workflowResultMessage = buildWorkflowResultContextMessage(input);
+  let continuationOptions = input.continuationOptions ?? null;
+
+  for (;;) {
+    if (!(await service.isWorkflowInvocationCurrent(input.workspaceId, input.runId))) {
+      log.debug("Skipping superseded workflow continuation", {
+        workspaceId: input.workspaceId,
+        runId: input.runId,
+      });
+      return;
+    }
+
+    continuationOptions ??= await service.getWorkflowContinuationSendOptions(input.workspaceId);
+    if (continuationOptions == null) {
+      log.warn("Skipping workflow continuation without send options", {
+        workspaceId: input.workspaceId,
+        runId: input.runId,
+      });
+      return;
+    }
+
+    const sendResult = await service.sendMessage(
+      input.workspaceId,
+      workflowResultMessage,
+      {
+        ...continuationOptions,
+        skipAiSettingsPersistence: true,
+        muxMetadata: {
+          type: WORKFLOW_RESULT_METADATA_TYPE,
+          rawCommand: input.rawCommand,
+          commandPrefix,
+          runId: input.runId,
+          requestedModel: continuationOptions.model,
+        },
+      },
+      {
+        skipAutoResumeReset: true,
+        synthetic: true,
+        agentInitiated: true,
+        requireIdle: true,
+        startStreamInBackground: true,
+      }
+    );
+    if (sendResult.success) {
+      return;
+    }
+    if (
+      sendResult.error.type !== "unknown" ||
+      sendResult.error.raw?.includes(WORKSPACE_BUSY_IDLE_ONLY_SEND_MESSAGE) !== true
+    ) {
+      log.warn("Failed to continue workflow after completion", {
+        workspaceId: input.workspaceId,
+        runId: input.runId,
+        error: sendResult.error,
+      });
+      return;
+    }
+    await new Promise((resolve) => setTimeout(resolve, WORKFLOW_CONTINUATION_RETRY_DELAY_MS));
+  }
 }
 
 export interface WorkspaceServiceEvents {
